requirements.txt
numpy==1.23.3
pybullet==3.2.5
PyYAML==6.0
